PARQUET-1819: [C++] Refactor decoding
Also add an additional size check before reading the length of a byte array.
